+++ contracts/registry/interface/IIdentityRegistry.sol
@@ -103,6 +103,12 @@ event IdentityUpdated(IIdentity indexed _oldIdentity, IIdentity indexed _newIden
 /// @param _country is the numeric code (ISO 3166-1) of the new country
 event CountryUpdated(address indexed _investorAddress, uint16 indexed _country);
 
+/// @dev This event is emitted when Eligibility checks are disabled.
+event EligibilityChecksDisabled();
+
+/// @dev This event is emitted when Eligibility checks are enabled.
+event EligibilityChecksEnabled();
+
 
 interface IIdentityRegistry {
 
@@ -191,6 +197,39 @@ interface IIdentityRegistry {
         IIdentity[] calldata _identities,
         uint16[] calldata _countries
     ) external;
+
+    /**
+     * @dev Disables the eligibility checks for token transfers and other operations.
+     *
+     * This function allows the token owner to disable the eligibility checks, effectively bypassing the
+     * complex verification process normally required for token operations. Once the eligibility checks
+     * are disabled, all users will be considered verified by the `isVerified` function, allowing them
+     * to interact with the token without requiring specific claims or issuer validation.
+     *
+     * Requirements:
+     * - The caller must be the owner of the contract.
+     * - The eligibility checks must not already be disabled; otherwise, the function will revert with an
+     *   `EligibilityChecksDisabledAlready` error.
+     *
+     * Emits an `EligibilityChecksDisabled` event upon successful execution.
+     */
+    function disableEligibilityChecks() external;
+
+    /**
+     * @dev Enables the eligibility checks for token transfers and other operations.
+     *
+     * This function allows the token owner to re-enable the eligibility checks after they have been
+     * disabled. Once re-enabled, the `isVerified` function will resume performing the full verification
+     * process, checking for the required claims and validating them through trusted issuers.
+     *
+     * Requirements:
+     * - The caller must be the owner of the contract.
+     * - The eligibility checks must currently be disabled; otherwise, the function will revert with an
+     *   `EligibilityChecksEnabledAlready` error.
+     *
+     * Emits an `EligibilityChecksEnabled` event upon successful execution.
+     */
+    function enableEligibilityChecks() external;
 
     /**
      *  @dev This functions checks whether a wallet has its Identity registered or not