+++ contracts/registry/storage/IRStorage.sol
@@ -77,9 +77,12 @@ contract IRStorage {
     /// @dev Address of the IdentityRegistryStorage Contract
     IIdentityRegistryStorage internal _tokenIdentityStorage;
 
+    /// @dev disables the whole eligibility check system
+    bool internal _checksDisabled;
+
     /**
      * @dev This empty reserved space is put in place to allow future versions to add new
      * variables without shifting down storage in the inheritance chain.
      */
-    uint256[49] private __gap;
+    uint256[48] private __gap;
 }